+++ vitest.config.ts
@@ -1,30 +1,36 @@
-import { aliasTs } from '@bemedev/vitest-alias';
-import { exclude } from '@bemedev/vitest-exclude';
+import { aliasTs } from '@bemedev/dev-utils/vitest-alias';
+import { exclude } from '@bemedev/dev-utils/vitest-exclude';
 import { defineConfig } from 'vitest/config';
 import tsconfig from './tsconfig.json';
 
 export default defineConfig({
+  server: {
+    host: '0.0.0.0',
+  },
   plugins: [
     aliasTs(tsconfig as any),
-    exclude({ ignoreCoverageFiles: ['**/index.ts', '**/types.ts'] }),
+    exclude({
+      ignoreCoverageFiles: [
+        '**/index.ts',
+        '**/types.ts',
+        '**/*.test-d.ts',
+      ],
+    }),
   ],
   test: {
-    bail: 10,
-    maxConcurrency: 10,
+    bail: 100,
+    maxConcurrency: 30,
     passWithNoTests: true,
     slowTestThreshold: 3000,
     globals: true,
     logHeapUsage: true,
     typecheck: {
       enabled: true,
-      only: false,
-      ignoreSourceErrors: true,
+      checker: 'tsc',
     },
     coverage: {
       enabled: true,
-      extension: 'ts',
       reportsDirectory: '.coverage',
-      all: true,
       provider: 'v8',
     },
   },